// sdk/src/common/types.ts
export interface DatasetDescriptor {
  id: string | number;
  name: string;
  label?: string;
  data_models: {
    id: string | number;
    name: string;
    label?: string;
    fields: {
      name: string;
      label?: string;
      type: string;
      is_custom_measure: boolean;
      defined_in_dataset?: boolean;
    }[];
  }[];
  metrics: {
    name: string;
    label?: string;
    type: string;
  }[];
}